ape: move compatibility libc.h to lib9 directory and incooperate needed functions for ape/mp (wip)

+++ sys/src/ape/lib/9/setmalloctag.c
@@ -1,4 +1,6 @@
+#include <stdint.h>
+
 void
-setmalloctag(void*, unsigned long)
+setmalloctag(void*, uintptr_t)
 {
 }
